Fix GitHub Actions workflows failing due to using glibc 2.17
PR #23 (#0774bd6) added a container setup to build with glibc 2.17. Unfortunately, GitHub recently started forcing actions to run on Node 20 instead of Node 16 (https://github.blog/changelog/2024-03-07-github-actions-all-actions-will-run-on-node20-instead-of-node16-by-default/). Node 20 does not work with glibc 2.17 and using it results in the build failing.

This commit temporarily fixes build issues by using the environment variable mentioned on the aforementioned page. We will need a longer term solution soon however, as once download-artifact@v3 and upload-artifact@v3 are deprecated, we will be forced to use v4 versions which require Node 20.