Use -quiet for all apt-get update usages and use consistent option ordering for apt-get
apt-get update is pretty noisy by default as there are a lot of lines for it reading the database. This should make it a bit quieter

I also noticed in the later part of the script -yq was before the install/update command so I made that consistent as well.